refactor: use Typescript
This is a relatively large commit to update the project to use Typescript.

The main benefit is documentation. By annotating the expected type for most values, it can be checked statically and help new maintainers to be confident with their changes. Beyond this, it brings the benefits of type definitions to consumers of the library (completion, type checking, etc).

This first commit does not do any change in the behavior of the code, it does not even type anything. It just adds Typescript and converts the files.

It adds `tslint` as a linter. This commit still results in lint errors but they will be fixed in a subsequent commit.

Even at this minimal stage, it caught a few errors:
- `JSON.parse` received a `Buffer` instead of a `string` in `parse-arg.js` line 8.
- In the integrations tests, it tried to strinfy the `output` field as if was the `stdout` buffer (but it is an array of buffers for the standard streams)
- `bin/c8.js` had to be executable when running the tests. This breaks on Windows

Closes bcoe/c8#17